resolve alias
- refactor types
- move static types in /entities
- create task to resolve alias
