+++ packages/concord/package.json
@@ -11,9 +11,9 @@
   },
   "keywords": [],
   "main": "dist/index.js",
-  "types": "dist/@types/index.d.ts",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "dev": "tsc && tsc-alias",
+    "dev": "gulp dev && gulp alias",
     "build": "gulp build && gulp header"
   },
   "dependencies": {
@@ -21,14 +21,12 @@
     "dotenv": "^10.0.0"
   },
   "devDependencies": {
-    "@types/node": "^16.11.11",
     "gulp": "^4.0.2",
     "gulp-header-comment": "^0.10.0",
-    "gulp-minify": "^3.1.0",
     "gulp-prettier": "^4.0.0",
     "gulp-rimraf": "^1.0.0",
     "gulp-typescript": "^6.0.0-alpha.1",
-    "tsc-alias": "^1.4.2",
+    "tsc-alias": "^1.5.0",
     "typescript": "^4.5.2"
   }
 }
