+++ packages/concord/src/entities/enums.ts
@@ -1,7 +1,3 @@
-export const enum commandsList {
-  ping = 'ping'
-}
-
 export const enum colorsList {
   ok = '#92D050',
   error = '#ED1C24'
