--- packages/concord/src/types.ts
@@ -1,5 +1,0 @@
-import { Intents } from 'discord.js'
-
-export const FLAGS = Intents.FLAGS
-
-export type TOptions = Partial<TOptionsDefault>
