+++ packages/testbot/src/index.js
@@ -1,5 +1,2 @@
 const { Bot } = require('@edixon/concord')
-
-const bot = new Bot()
-
-bot.start()
+new Bot().start()
